core: ensure custom/runtime metrics configured similarly (#3590)
### What does this PR do?
- refactor metric config logic into a common location
- pass this config to both runtime and custom metrics

### Motivation
- runtime metrics and custom metrics should always be sent to the same place
- probably nobody will want them sent to multiple locations
- currently this can result in apps where RM are received but CM are dropped
